refactor: move models to models  folder
