+++ src/routes/graphql/models/member-type.model.ts
@@ -5,15 +5,15 @@ import {
   GraphQLNonNull,
   GraphQLList,
 } from 'graphql';
-import { MemberTypeIdType } from './member-type-id.js';
-import { ProfileType } from './profile-type.js';
+import { MemberTypeIdType } from '../types/member-type-id.js';
+import { Profile } from './profile.model.js';
 
-export const MemberTypeType = new GraphQLObjectType({
+export const MemberType = new GraphQLObjectType({
   name: 'MemberType',
   fields: () => ({
     id: { type: new GraphQLNonNull(MemberTypeIdType) },
     discount: { type: new GraphQLNonNull(GraphQLFloat) },
     postsLimitPerMonth: { type: new GraphQLNonNull(GraphQLInt) },
-    profiles: { type: new GraphQLList(ProfileType) },
+    profiles: { type: new GraphQLList(Profile) },
   }),
 });
