+++ src/routes/graphql/models/profile.model.ts
@@ -1,9 +1,9 @@
 import { GraphQLObjectType, GraphQLBoolean, GraphQLInt, GraphQLNonNull } from 'graphql';
-import { UUIDType } from './uuid.js';
-import { MemberTypeType } from './member-type-type.js';
-import { MemberTypeIdType } from './member-type-id.js';
-import { UserType } from './user-type.js';
-import { ContextValueType } from './context-value-type.js';
+import { UUIDType } from '../types/uuid.js';
+import { MemberType } from './member-type.model.js';
+import { MemberTypeIdType } from '../types/member-type-id.js';
+import { User } from './user.model.js';
+import { ContextValueType } from '../types/context-value-type.js';
 
 const resolve = async (
   source: { memberTypeId: string },
@@ -17,16 +17,16 @@ const resolve = async (
   });
 };
 
-export const ProfileType: GraphQLObjectType = new GraphQLObjectType({
+export const Profile: GraphQLObjectType = new GraphQLObjectType({
   name: 'Profile',
   fields: () => ({
     id: { type: new GraphQLNonNull(UUIDType) },
     isMale: { type: new GraphQLNonNull(GraphQLBoolean) },
     yearOfBirth: { type: new GraphQLNonNull(GraphQLInt) },
 
-    user: { type: UserType },
+    user: { type: User },
     userId: { type: new GraphQLNonNull(UUIDType) },
-    memberType: { type: new GraphQLNonNull(MemberTypeType), resolve },
+    memberType: { type: new GraphQLNonNull(MemberType), resolve },
     memberTypeId: { type: new GraphQLNonNull(MemberTypeIdType) },
   }),
 });
