+++ src/routes/graphql/mutations/change-post.ts
@@ -1,12 +1,12 @@
 import { GraphQLNonNull } from 'graphql';
 import { ContextValueType } from '../types/context-value-type.js';
-import { PostType } from '../types/post-type.js';
+import { Post } from '../models/post.model.js';
 import { UUIDType } from '../types/uuid.js';
 import { ChangePostInputType } from '../types/change-post-input-type.js';
 import { PostInput } from '../types/input-types.js';
 
 export const changePost = {
-  type: PostType,
+  type: Post,
   args: {
     id: { type: new GraphQLNonNull(UUIDType) },
     dto: { type: new GraphQLNonNull(ChangePostInputType) },
