+++ src/routes/graphql/mutations/create-profile.ts
@@ -1,11 +1,11 @@
 import { GraphQLNonNull } from 'graphql';
 import { ContextValueType } from '../types/context-value-type.js';
-import { ProfileType } from '../types/profile-type.js';
+import { Profile } from '../models/profile.model.js';
 import { CreateProfileInputType } from '../types/create-profile-input-type.js';
 import { ProfileInput } from '../types/input-types.js';
 
 export const createProfile = {
-  type: ProfileType,
+  type: Profile,
   args: { dto: { type: new GraphQLNonNull(CreateProfileInputType) } },
   resolve: async (_, args: { dto: ProfileInput }, context: ContextValueType) => {
     return await context.prisma.profile.create({
