+++ src/routes/graphql/queries/member-types-query.ts
@@ -1,9 +1,9 @@
 import { GraphQLList, GraphQLResolveInfo } from 'graphql';
 import { ContextValueType } from '../types/context-value-type.js';
-import { MemberTypeType } from '../types/member-type-type.js';
+import { MemberType } from '../models/member-type.model.js';
 
 export const memberTypesQuery = {
-  type: new GraphQLList(MemberTypeType),
+  type: new GraphQLList(MemberType),
   resolve: async (_s, _a, context: ContextValueType, _i: GraphQLResolveInfo) => {
     return await context.prisma.memberType.findMany();
   },
