+++ src/routes/graphql/queries/user-query.ts
@@ -1,5 +1,5 @@
 import { ContextValueType } from '../types/context-value-type.js';
-import { UserType } from '../types/user-type.js';
+import { User } from '../models/user.model.js';
 import { UUIDType } from '../types/uuid.js';
 
 const resolve = async (_, args: { id: string }, context: ContextValueType) => {
@@ -7,7 +7,7 @@ const resolve = async (_, args: { id: string }, context: ContextValueType) => {
 };
 
 export const userQuery = {
-  type: UserType,
+  type: User,
   args: { id: { type: UUIDType } },
   resolve,
 };
